README: mention SLES12 SP2 Beta in history
In june 2016 we have already given an smc-tools version to SuSE
for inclusion into SLES12 SP2. It matched the SMC-R kernel code
of june 2016. It made it into some Beta-Version of SLES12 SP2.
Even though this version 1.0.0 is quite different, we should
mention this SLES12 SP2 Beta version in the history part of
README.smc-tools.

Signed-off-by: Ursula Braun <ubraun@linux.vnet.ibm.com>