Source version number from a single place
Convert README, build_tar, smc-tools.spec and af_smc.7 to retrieve the version
number from Makefile, so there is only a single place that stores the version
number.
Note that smc-tools.spec requires some special handling in the Makefile as it
is a generated file that is not shipped in the .tgz and hence must not appear
as a dependency in the 'all' target - yet still needs to be built.
While at it, fix a typo in Makefile, and add some minor cleanups to build_tar.

Signed-off-by: Stefan Raspl <raspl@linux.vnet.ibm.com>