Makefile: Fix version information
VERSION, RELEASE and PATCHLEVEL are unused apart from SMC_TOOLS_RELEASE,
removing.
Furthermore, build-tar sanity-checks for a README entry for version
SMC_TOOLS_RELEASE, but would currently fail as the README specifies the
next version as 1.x.x, while Makefile uses 1.0.1. Since we cannot know
at this point what will be the next version number, converting to 1.x.x
all the way. This has the added benefit of making it really obvious that
this is a development version, which should help avoid confusion due to
incorrectly claiming issues to be found in a released version when
reporting bugs.