build-tar: Fix messed up 'install' target in Makefile
The .tgz created by build-tar will fail when 'make install' is called.
This is due to the excess removal of af_smc.7 from the install target when
creating the .tgz.
The fix is a bit ugly: There's 3 occurences of af_smc.7, and only the first
and the last need to be removed. That's why the pattern we're matching is
what it is, to do exactly that: Leave the middle occurence in place.

Signed-off-by: Stefan Raspl <raspl@linux.ibm.com>