smcss: display fallback reason to user
Introduce a new extension FALLBACK that holds the fallback reason and
an optional peer diagnosis code.
And check the length of each extension before accessing it, in case an
extension is enlarged later and older kernels still send the short
version.
And rename the --extended parameter into --debug, update man page
accordingly.

Signed-off-by: Karsten Graul <kgraul@linux.ibm.com>
Reviewed-by: Ursula Braun <ubraun@linux.ibm.com>