smc-tools: rename preload library
The preload library is renamed into libsmc-preload.so. The former name
did not follow the standard library name schema (lib*.so*, see FILES
section of 'man ld.so'), so the library was not processed and made
available in a standard way and the smc_run script had to do some work
to enable to library. And add a %post step to the spec file to call
ldconfig after rpm installation so the new library becomes active.

Signed-off-by: Karsten Graul <kgraul@linux.ibm.com>