rpm: Adapt Fedora's .spec file
Fedora is packaging smc-tools, and we're conveniently lifting their spec
file.
Note that we have to turn off 32-but .so installsin the Makefile:
rpmbuild apparently doesn't support a single package installing code for
multiple architectures. And since rpmbuild does all kinds of sanity-check,
it will complain in case the Makefile's 'install' target installs a 32-bit
library, but the .spec file doesn't package it.

Signed-off-by: Stefan Raspl <raspl@linux.ibm.com>