smc_rnics: Misc fixes
Turns out the PNET ID was not correctly interpreted (broken by a previous
patch). Also, the ports were not always correctly display. Finally, there
is some unresolvable issue in case a one of the modules mlx4_ib and mlx_en
is unloaded: This makes directories disappear. And since we're relying on
both directories to be present, we'd need quite a bit of confusing add'l
code to handle these cases. Which I would argue are quite exotic. And
besides: Somebody who unloads modules of the device driver cannot expect
things to still flow perfectly.

Signed-off-by: Stefan Raspl <raspl@linux.ibm.com>