smc_rnics: Suppress unknown non-networking devices
As reported by Test, non-networking devices are still visible (see last
line):
  $ smc_rnics
  FID  Power  PCI ID        PCHID  Type           Port  PNET ID   Interface
  -------------------------------------------------------------------------
    0  1      0000:00:00.0  0150   RoCE Express2  0     PERFNET   enp0s0
   40  1      000e:00:00.0  011c   RoCE Express2  0     PERFNET   enP14s40
   80  1      000d:00:00.0  013c   RoCE Express2  0     PERFNET   enP13s80
   c0  1      000c:00:00.0  0170   RoCE Express2  0     PERFNET   enP12p0s0
  800  1      0004:00:00.0  0110   144d:a822
To address this, we are now defaulting to supression of any unknown device
in the default output. However, if option '-r' is specified, we still show
all output, including non-networking devices - this can be useful in case
an older version is run with later hardware.
Note that it is tedious to blacklist all non-networking devices: Since
users can add their own NVMe devices, we can never know what we will
encounterin the wild.

Signed-off-by: Stefan Raspl <raspl@linux.ibm.com>