Adapt smc program to create 2 separate programs for smcd and smcr
The smc program offers the type parameter to display smcd or smcr
related output. The default is smcr, but this way the output of the
smc command without any parameters on an smcd-only system shows simply
nothing, which may be confusing for users.
For now we decided to split up the functionality into 2 programs smcd
and smcr that are build out of the original smc program source code,
instrumented with preprocessor commands.

This way we keep the way open for the introduction of the smc program
that then should offer support for both types of SMC together.

Signed-off-by: Karsten Graul <kgraul@linux.ibm.com>
Signed-off-by: Guvenc Gulce <guvenc@linux.ibm.com>