smc_tools: fix PNET ID for multiple port devices
The util_string for a CHPID is a 64-byte field with up to 4 16-bytes
PNET IDs belonging to up to 4 ports of the CHPID.
There is no need to define a PNET ID for every CHPID port. If undefined
the PNETID value consists of 16 bytes equal to hex 00.

This patch enables the smc_tools commands smc_rnics and smc_dbg to
deal with multiple port CHPIDs and their mix of undefined and defined
PNET IDs in their corresponding util_strings.
It adds a "Port" column to "CCW Devices", since different ports usually
have different PNET IDs.

Signed-off-by: Ursula Braun <ubraun@linux.ibm.com>