smc_run: Add various command-line switches
Add switches to set send and receive buffer-sizes, as well as the universal
'-h' and '-v'.
This requires changes in the preload library. Values are passed on via env
variables SMC_SNDBUF and SMC_RCVBUF, and can be specified in Bytes either
plain, or using metric prefixes k/K and m/M. Which helps to take advantage
of this feature in absence of using smc_run.
Furthermore, we're limited by net.core.rmem_max and net.core.wmem_max, which
are set to 256k by default. Therefore, we raise those values if they are
less than what the user requests. This is documented in the man page.
While at it, reworked (simplified) the opts parsing logic a bit.

Signed-off-by: Stefan Raspl <raspl@linux.ibm.com>