smc_chk: Fix UEID calling convention
smcd/r supported the "eid" command only briefly during development, before
it was converted to "ueid". However, we never fixed this in this currently
unused section of code.
Since users would have to study the code to reproduce (the respective
commandline switches are missing from online help and man page), we don't
document this glitch.

Signed-off-by: Stefan Raspl <raspl@linux.ibm.com>