Makefile: Fix target 'check'
We were still trying to test the 'smc' command, which was converted into
'smcd' and 'smcr' a while ago. We change this to test 'smcd info' and
'smcd stats' instead for basic coverage. Note: Both of these commands are
effectively identical between 'smcd' and 'smcr', so testing either one is
good enough.
While at it, we also convert the 'which' calls to something more
standard-abiding.

Signed-off-by: Stefan Raspl <raspl@linux.ibm.com>