adjust default library path for Ubuntu
For whatever reason the default library path for Ubuntu is
lib/${ARCHTYPE}-linux-gnu in smc-tools (which expands to lib/s390-linux-gnu), but
for other distributions it's /lib64, which translates to /lib in Ubuntu.
This changes the default patch on Ubuntu from lib/${ARCHTYPE}-linux-gnu
to (standard) /lib which is comparable to /lib64 on other distros.
On Ubuntu /lib holds by default 64-bit libraries.

Signed-off-by: Jan Karcher <jaka@linux.ibm.com>