util: fix build with CFLAGS="-g"
The build process fails with debug symbols enabled. It appears that
`CFLAGS=-g` influences the compiler's inlining decisions, leading to a
situation where the `is_str_empty` function is not inlined, and the
linker is searching for an out-of-line definition that isn’t present.
The most straightforward fix is to declare this function
`static inline`.

Signed-off-by: Wolfgang Frisch <wfrisch@suse.de>